Correct boxplot and MS-plot
