+++ skfda/_utils/__init__.py
@@ -1,5 +1,4 @@
 from . import constants
 
 from ._utils import (_list_of_arrays, _coordinate_list,
-                     _check_estimator, parameter_aliases,
-                     _create_figure, _figure_to_svg)
+                     _check_estimator, parameter_aliases)
