+++ skfda/_utils/_utils.py
@@ -1,12 +1,9 @@
 """Module with generic methods"""
 
 import functools
-import io
+
 import types
 
-import matplotlib.backends.backend_svg
-
-import matplotlib.pyplot as plt
 import numpy as np
 
 
@@ -141,22 +138,3 @@ def _check_estimator(estimator):
     instance = estimator()
     check_get_params_invariance(name, instance)
     check_set_params(name, instance)
-
-
-def _create_figure():
-    """Create figure using the default backend."""
-    fig = plt.figure()
-
-    return fig
-
-
-def _figure_to_svg(figure):
-    """Return the SVG representation of a figure."""
-
-    old_canvas = figure.canvas
-    matplotlib.backends.backend_svg.FigureCanvas(figure)
-    output = io.BytesIO()
-    figure.savefig(output, format='svg')
-    figure.set_canvas(old_canvas)
-    data = output.getvalue()
-    return data.decode('utf-8')
