+++ skfda/exploratory/visualization/_magnitude_shape_plot.py
@@ -6,19 +6,14 @@ detection method is implemented.
 
 """
 
-from io import BytesIO
-
 import matplotlib
-from scipy.stats import f, variation
-from sklearn.covariance import MinCovDet
 
 import matplotlib.pyplot as plt
 import numpy as np
 
-from ..._utils import _create_figure, _figure_to_svg
 from ..depth import modified_band_depth
-from ..outliers import (directional_outlyingness_stats,
-                        DirectionalOutlierDetector)
+from ..outliers import DirectionalOutlierDetector
+from ._utils import _figure_to_svg, _get_figure_and_axes, _set_figure_layout
 
 
 __author__ = "Amanda Hernando Bernabé"
@@ -241,7 +236,7 @@ class MagnitudeShapePlot:
                 "outcol must be a number between 0 and 1.")
         self._outliercol = value
 
-    def plot(self, ax=None):
+    def plot(self, chart=None, *, fig=None, axes=None,):
         """Visualization of the magnitude shape plot of the fdatagrid.
 
         Args:
@@ -252,23 +247,23 @@ class MagnitudeShapePlot:
             fig (figure object): figure object in which the graph is plotted.
 
         """
+
+        fig, axes = _get_figure_and_axes(chart, fig, axes)
+        fig, axes = _set_figure_layout(fig, axes)
+
         colors = np.zeros((self.fdatagrid.n_samples, 4))
         colors[np.where(self.outliers == 1)] = self.colormap(self.outliercol)
         colors[np.where(self.outliers == 0)] = self.colormap(self.color)
 
-        if ax is None:
-            fig = _create_figure()
-            ax = fig.add_subplot(1, 1, 1)
-
         colors_rgba = [tuple(i) for i in colors]
-        ax.scatter(self.points[:, 0].ravel(), self.points[:, 1].ravel(),
-                   color=colors_rgba)
+        axes[0].scatter(self.points[:, 0].ravel(), self.points[:, 1].ravel(),
+                        color=colors_rgba)
 
-        ax.set_xlabel(self.xlabel)
-        ax.set_ylabel(self.ylabel)
-        ax.set_title(self.title)
+        axes[0].set_xlabel(self.xlabel)
+        axes[0].set_ylabel(self.ylabel)
+        axes[0].set_title(self.title)
 
-        return ax.get_figure()
+        return fig
 
     def __repr__(self):
         """Return repr(self)."""
